Updated this analyzer to use domains instead of hosts
The issue with using hosts with the latest Cuckoo is that the hosts data only contains IP addresses and no domains or countries like it did in the past. Instead we can use the domains data parameter to pull in IP and Domain which is much more beneficial then the reports today. We will need to update the report template as well.