StamusNetworks: add tenant support
Scirius Security Platform supports multitenant where multiple
organization network can use the same platform. If this feature
is activated then the tenant option with value being the one of
the organization has to be used to get access to the correct data.

This matches the organization approach of Cortex and TheHive.