New Analyzer & Responders Watcher
This PR introduces two new tools for Cortex related to the Watcher - Open Source Cybersecurity Threat Hunting Platform.

1. Watcher Check Domain Analyzer:

This analyzer allows TheHive/Cortex to check if a given domain is already being monitored by Watcher. Its main feature is to quickly verify the monitoring status of a domain.

2. Watcher Monitor Manager Responder:

This responder enables TheHive/Cortex to manage the monitoring of domains within Watcher. It provides two main features:
    • Add a domain to monitoring (WatcherAddDomain)
    • Remove a domain from monitoring (WatcherRemoveDomain)

These tools together provide a way to both check the current monitoring status of a domain in Watcher and to take action to add or remove a domain from monitoring directly from TheHive/Cortex.